extracted learning params from single agent QL and trying if gym env registration works
