+++ __init__.py
@@ -1,0 +1,8 @@
+from gym.envs.registration import register
+
+register(
+    id='unimore/SlimeEnvSingleAgent-v0',
+    entry_point='environments/SlimeEnvSingleAgent',
+    max_episode_steps=10000,  # DOC The keyword argument max_episode_steps=300 will ensure that GridWorld environments that are instantiated via gym.make will be wrapped in a TimeLimit wrapper (https://www.gymlibrary.dev/content/environment_creation/#registering-envs)
+    nondeterministic=True  # DOC seeding not supported atm
+)
