+++ agents/sa-learning-params.json
@@ -1,0 +1,12 @@
+{
+  "alpha": 0.2,
+  "gamma": 0.8,
+  "epsilon": 0.9,
+  "decay": 0.9995,
+  "train_episodes": 2,
+  "TRAIN_LOG_EVERY": 1,
+  "test_episodes": 2,
+  "TEST_LOG_EVERY": 1,
+  "OUTPUT_FILE": "single-test-01",
+  "actions": ["move-toward-chemical", "random-walk", "drop-chemical"]
+}
