+++ environments/__init__.py
@@ -1,0 +1,1 @@
+from environments.SlimeEnvSingleAgent import Slime
